SQL file to create tables from external database
This is the file to create the tables from the external database in our database. It creates the following tables: user, user_login, cart, and goes_in. It also adds a sales column to the product tables so that we can calculate sales. Finally, it also adds the following views: product_sales and total_sales. These views display the sales for individual products and the total sales of all the products respectively.